backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
jinja2
pdfkit
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform
google-generativeai
python-dotenv


